requirements.txt
chardet
fpdf2
progress
